# openapi.json
{"openapi":"3.0.1","info":{"title":"World at War: Turn-Based Strategy API","description":"API for World at War - competitive turn-based strategy game","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Games","description":"API endpoints for game management"},{"name":"Authentication","description":"API endpoints for user authentication and profile management"},{"name":"Moves","description":"API endpoints for game moves management"}],"paths":{"/api/games/{gameId}/status":{"put":{"tags":["Games"],"summary":"Update game status","description":"Updates the status of a game (surrender, abandon)","operationId":"updateGameStatus","parameters":[{"name":"gameId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGameStatusRequest"}}},"required":true},"responses":{"200":{"description":"Game status updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UpdateGameStatusResponse"}}}},"403":{"description":"You are not a participant or invalid status transition","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UpdateGameStatusResponse"}}}},"404":{"description":"Game not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UpdateGameStatusResponse"}}}},"422":{"description":"Invalid status transition","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UpdateGameStatusResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/games":{"get":{"tags":["Games"],"summary":"Get games list","description":"Retrieves a paginated list of games for the current user with filtering","operationId":"getGames","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"gameType","in":"query","required":false,"schema":{"type":"string"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"Games list retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GameListResponse"}}}},"400":{"description":"Invalid query parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GameListResponse"}}}},"401":{"description":"Unauthorized - authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GameListResponse"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Games"],"summary":"Create new game","description":"Creates a new game (vs_bot or pvp)","operationId":"createGame","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGameRequest"}}},"required":true},"responses":{"201":{"description":"Game created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateGameResponse"}}}},"400":{"description":"Invalid game parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateGameResponse"}}}},"422":{"description":"Validation errors","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateGameResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/games/{gameId}/moves":{"get":{"tags":["Moves"],"summary":"Get moves history","description":"Retrieves all moves for a specific game","operationId":"getMoves","parameters":[{"name":"gameId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"You are not a participant of this game","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoveListItem"}}}}},"404":{"description":"Game not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoveListItem"}}}}},"200":{"description":"Moves retrieved successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoveListItem"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Moves"],"summary":"Create move","description":"Creates a new move in a game","operationId":"createMove","parameters":[{"name":"gameId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMoveRequest"}}},"required":true},"responses":{"201":{"description":"Move created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateMoveResponse"}}}},"404":{"description":"Game not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateMoveResponse"}}}},"400":{"description":"Game is not in progress","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateMoveResponse"}}}},"422":{"description":"Invalid move","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateMoveResponse"}}}},"403":{"description":"You are not a participant or not the current player","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateMoveResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/games/{gameId}/bot-move":{"post":{"tags":["Moves"],"summary":"Create bot move","description":"Creates a bot move in a vs_bot game (internal endpoint)","operationId":"createBotMove","parameters":[{"name":"gameId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Only player1 can trigger bot moves","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BotMoveResponse"}}}},"400":{"description":"Game is not vs_bot or not in progress","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BotMoveResponse"}}}},"404":{"description":"Game not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BotMoveResponse"}}}},"200":{"description":"Bot move created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BotMoveResponse"}}}},"500":{"description":"Failed to generate bot move","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BotMoveResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register new user","description":"Creates a new user account and returns JWT token","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"User successfully created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}},"422":{"description":"Validation error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}},"400":{"description":"Email or username already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}}}}},"/api/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout user","description":"Logs out the currently authenticated user and returns a confirmation message","operationId":"logout","responses":{"401":{"description":"Unauthorized - authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LogoutResponse"}}}},"200":{"description":"User logged out successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LogoutResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login user","description":"Authenticates a user and returns JWT token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successfully logged in","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"400":{"description":"Validation error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"401":{"description":"Invalid credentials","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/games/{gameId}":{"get":{"tags":["Games"],"summary":"Get game details","description":"Retrieves detailed information about a game including board state and move history","operationId":"getGameDetail","parameters":[{"name":"gameId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"You are not a participant of this game","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GameDetailResponse"}}}},"200":{"description":"Game details retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GameDetailResponse"}}}},"404":{"description":"Game not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GameDetailResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/games/{gameId}/board":{"get":{"tags":["Games"],"summary":"Get board state","description":"Retrieves the current board state for a game","operationId":"getBoardState","parameters":[{"name":"gameId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"You are not a participant of this game","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BoardStateResponse"}}}},"404":{"description":"Game not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BoardStateResponse"}}}},"200":{"description":"Board state retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BoardStateResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/auth/me":{"get":{"tags":["Authentication"],"summary":"Get current user profile","description":"Retrieves the profile of the currently authenticated user including statistics and metadata","operationId":"getCurrentUserProfile","responses":{"401":{"description":"Unauthorized - authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"200":{"description":"User profile retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"UpdateGameStatusRequest":{"required":["status"],"type":"object","properties":{"status":{"type":"string","enum":["waiting","in_progress","finished","abandoned","draw"]}}},"UpdateGameStatusResponse":{"type":"object","properties":{"gameId":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["waiting","in_progress","finished","abandoned","draw"]},"updatedAt":{"type":"string","format":"date-time"}}},"CreateGameRequest":{"required":["boardSize","gameType"],"type":"object","properties":{"gameType":{"type":"string","enum":["vs_bot","pvp"]},"boardSize":{"maximum":5,"minimum":3,"type":"integer","format":"int32"},"botDifficulty":{"type":"string","enum":["easy","medium","hard"]}}},"BoardState":{"type":"object","properties":{"state":{"type":"array","items":{"type":"array","items":{"type":"string"}}}}},"CreateGameResponse":{"type":"object","properties":{"gameId":{"type":"integer","format":"int64"},"gameType":{"type":"string","enum":["vs_bot","pvp"]},"boardSize":{"type":"string","enum":["THREE","FOUR","FIVE"]},"player1Id":{"type":"integer","format":"int64"},"player2Id":{"type":"integer","format":"int64"},"botDifficulty":{"type":"string","enum":["easy","medium","hard"]},"status":{"type":"string","enum":["waiting","in_progress","finished","abandoned","draw"]},"currentPlayerSymbol":{"type":"string","enum":["x","o"]},"createdAt":{"type":"string","format":"date-time"},"boardState":{"$ref":"#/components/schemas/BoardState"}}},"CreateMoveRequest":{"required":["playerSymbol"],"type":"object","properties":{"row":{"minimum":0,"type":"integer","format":"int32"},"col":{"minimum":0,"type":"integer","format":"int32"},"playerSymbol":{"type":"string","enum":["x","o"]}}},"CreateMoveResponse":{"type":"object","properties":{"moveId":{"type":"integer","format":"int64"},"gameId":{"type":"integer","format":"int64"},"row":{"type":"integer","format":"int32"},"col":{"type":"integer","format":"int32"},"playerSymbol":{"type":"string","enum":["x","o"]},"moveOrder":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"boardState":{"$ref":"#/components/schemas/BoardState"},"gameStatus":{"type":"string","enum":["waiting","in_progress","finished","abandoned","draw"]},"winner":{"$ref":"#/components/schemas/WinnerInfo"}}},"WinnerInfo":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"username":{"type":"string"}}},"BotMoveResponse":{"type":"object","properties":{"moveId":{"type":"integer","format":"int64"},"gameId":{"type":"integer","format":"int64"},"row":{"type":"integer","format":"int32"},"col":{"type":"integer","format":"int32"},"playerSymbol":{"type":"string","enum":["x","o"]},"moveOrder":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"boardState":{"$ref":"#/components/schemas/BoardState"},"gameStatus":{"type":"string","enum":["waiting","in_progress","finished","abandoned","draw"]},"winner":{"$ref":"#/components/schemas/WinnerInfo"}}},"RegisterRequest":{"required":["email","password","username"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"},"username":{"maxLength":50,"minLength":3,"type":"string"}}},"RegisterResponse":{"type":"object","properties":{"userId":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"isGuest":{"type":"boolean"},"totalPoints":{"type":"integer","format":"int64"},"gamesPlayed":{"type":"integer","format":"int32"},"gamesWon":{"type":"integer","format":"int32"},"authToken":{"type":"string"}}},"LogoutResponse":{"type":"object","properties":{"message":{"type":"string"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"userId":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"isGuest":{"type":"boolean"},"totalPoints":{"type":"integer","format":"int64"},"gamesPlayed":{"type":"integer","format":"int32"},"gamesWon":{"type":"integer","format":"int32"},"authToken":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"GameListItem":{"type":"object","properties":{"gameId":{"type":"integer","format":"int64"},"gameType":{"type":"string","enum":["vs_bot","pvp"]},"boardSize":{"type":"string","enum":["THREE","FOUR","FIVE"]},"status":{"type":"string","enum":["waiting","in_progress","finished","abandoned","draw"]},"player1Username":{"type":"string"},"player2Username":{"type":"string"},"winnerUsername":{"type":"string"},"botDifficulty":{"type":"string","enum":["easy","medium","hard"]},"totalMoves":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"lastMoveAt":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"}}},"GameListResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/GameListItem"}},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"GameDetailResponse":{"type":"object","properties":{"gameId":{"type":"integer","format":"int64"},"gameType":{"type":"string","enum":["vs_bot","pvp"]},"boardSize":{"type":"string","enum":["THREE","FOUR","FIVE"]},"player1":{"$ref":"#/components/schemas/PlayerInfo"},"player2":{"$ref":"#/components/schemas/PlayerInfo"},"winner":{"$ref":"#/components/schemas/WinnerInfo"},"botDifficulty":{"type":"string","enum":["easy","medium","hard"]},"status":{"type":"string","enum":["waiting","in_progress","finished","abandoned","draw"]},"currentPlayerSymbol":{"type":"string","enum":["x","o"]},"lastMoveAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"},"boardState":{"$ref":"#/components/schemas/BoardState"},"totalMoves":{"type":"integer","format":"int32"},"moves":{"type":"array","items":{"$ref":"#/components/schemas/MoveListItem"}}}},"MoveListItem":{"type":"object","properties":{"moveId":{"type":"integer","format":"int64"},"row":{"type":"integer","format":"int32"},"col":{"type":"integer","format":"int32"},"playerSymbol":{"type":"string","enum":["x","o"]},"moveOrder":{"type":"integer","format":"int32"},"playerId":{"type":"integer","format":"int64"},"playerUsername":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"PlayerInfo":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"username":{"type":"string"},"isGuest":{"type":"boolean"}}},"BoardStateResponse":{"type":"object","properties":{"boardState":{"$ref":"#/components/schemas/BoardState"},"boardSize":{"type":"string","enum":["THREE","FOUR","FIVE"]},"totalMoves":{"type":"integer","format":"int32"},"lastMove":{"$ref":"#/components/schemas/LastMove"}}},"LastMove":{"type":"object","properties":{"row":{"type":"integer","format":"int32"},"col":{"type":"integer","format":"int32"},"playerSymbol":{"type":"string","enum":["x","o"]},"moveOrder":{"type":"integer","format":"int32"}}},"UserProfileResponse":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"username":{"type":"string"},"isGuest":{"type":"boolean"},"totalPoints":{"type":"integer","format":"int64"},"gamesPlayed":{"type":"integer","format":"int32"},"gamesWon":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"lastSeenAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT authorization token","scheme":"bearer","bearerFormat":"JWT"}}}}